servers/rest_firstchat_api/tests/requirements.txt
streamlit>=1.22.0
pillow>=9.0.0
requests>=2.28.0